Consolidate tests and drop all asterisks
Testing code is centralized even more, since everything is in the same module now.

Testing code never use the attribute for ReactiveForm and instead set the the
validator on the (external) formControl as we expect to do in a real-world
use-case.

`[required]` is never used internally anymore. Because internally we always use
a formControl and thus we must not interfer with a template-driven attribute.
That means we lost the asterisk, in all use-cases this time, but it will be
restored in a later commit.