New `ifValid` to handle async validated forms #7496
A form might have async validators. If a form is pending, the form
submission should wait for valid status and then be submitted (instead
of giving up the submission altogether).

This new function wraps that logic in a fool-proof way and should be used
everywhere we check for form validity (except in templates)