Support for NavigableItem in mass deletion #7910
Selection management via `masterToggle()` and `isAllSelected()` is strictly
available for the standard `dataSource`. Other use-cases should be handled
separately to guarantee consistent, strong typing and ease of use for normal
cases.

Also NaturalAbstractList now properly support that `dataSource` might be
navigable or not and thus handle the selection accordingly, always keeping
non-navigable item in the selection.