Let the browser decide how to present file #7880
In most cases browsers will show PDF without prompting to download.
This is the most common use-case. And users are still able to download
the document by clicking the browser specific button after seeing the
preview.

Also this attribute prevent the server to maybe suggest something else,
so it's best to avoid it. And if really needed rely on something the
server instead.