BREAKING refactor columns persistence #7949
We cannot rely on `this.route.snapshot.data.initialColumns` in
`selectColumns()` because the component might not be routed, but
instead template-d. So the data in routing is not for him, and
instead we should rely on the direct input in `this.initialColumns`.

That means that AbstractList must have a way to set the column selection
after initialization. This now done with two way binding. Simplifying
usage (and thus breaking the API).

This also avoid forgetting that a column might be available but unchecked
via template and thus not part of default selection.

We exhaustively covered all cases with tests, which allow us to safely
simplify the persistent format for shorter and prettier URL.

Slightly out of scope, but an even better thing would be to never call
`initFromRoute` in a template-d component. But I don't know of an automatic
way to tell if the component is routed or templated. So for now we can
workaround the mixing up of route data and direct inputs by always specifying
an `[initialColumns]` if an AbstractList is templated withing another component
that is routed with some `initialColumns` route data. The same idea applies to
`forcedVariables`.