Allow 404 fallback mechanism with panels #7909
Angular routing uses `path: '**'` as fallback mechanism to catch all
404 and do something more useful than crash and burn.

With our panel panel system this was not possible anymore because
panels require to catch Angular navigation errors (to re-route them inside
panels) and thus adding `path: '**'` broke panel navigations.

So we introduce `matcher: fallbackIfNoOpenedPanels` as a replacement of
`path: '**'`. This allows Angular to fallback as usual when panels are not
opened. But as soon as at least one panel is opened, then Angular fallback
is not matched anymore and our panel routing takes over.

So the panel logic is almost untouched, but we "enable" it only when necessary.