BREAKING Cancel update mutation when doing delete mutation #8110
After we deleted an object we cannot update it after the debounce
anymore. So we cancel the entire debounce mechanism upon object
deletion.

This also migrate away from lodash to rxjs operators to better
integrate with observable. Most importantly we previously had
all debounced-so-ignored update observables never complete. This
would lead to potentially severe memory leak in the whole app since
it is a very frequent occurrence to have a user type and pause and
type again.

It is a (minor) breaking change because `getKey` was dropped entirely.
But otherwise the general behavior should not be breaking.