Exclude the icon name from Google search result #8255
The native `<mat-icon>` should never be used anymore.

See https://developers.google.com/search/docs/advanced/crawling/special-tags#data-nosnippet

This should work on span, div an section, but not sure it would work
on a custom element. So this is a test. If it work, then it can be
generalized for all natural-icon (in Natural) and mat-icon (in Theodia)