Prevent double-deletion #9074
Form starts to be disabled before showing confirmation, so that is is no
longer possible to re-click on the delete button just after confirming
dialog but before the form is disabled.

We also keep the form disabled until after the redirect, to be sure that
we don't update an object that has just been deleted.