Can re-use components event when XHR failed in the past #9221
When Angular reuse components that means we also reuse our observables
and their existing subscriptions, typically the
`NaturalAbstractModelService.watchAll()`. But if that observable failed
in the past, eg: because of an intermittent network error, then the
observable would have errored and died forever. That means we can never
receive any data as long as we reuse that same component.

To avoid all of this we need to keep the observable alive forever, even
if there are errors. So errors in `watchAll` will be totally silenced.
But, typically, our app is configured to see errors at a lower level, in
the Apollo Link, so the app can still show a snackbar with the error to
end-user.