Cannot cumulate any `mat-*button` on the same elements #9432
Especially `mat-icon-button` will crash at runtime if used in
combination with any other `mat-*button`.

That also means that out of the box we cannot have a raised
`mat-icon-button`. So we reproduce the result by applying the new custom
classes such as `icon-button-circle-accent` for icon with color, or
alternatively the existing class `mat-elevation-z4` for an icon without
color.